Create separate accounts app and move team management there
