+++ localshop/apps/accounts/urls.py
@@ -1,10 +1,9 @@
 from django.conf.urls import url
 
-from localshop.apps.permissions import views
+from localshop.apps.accounts import views
 
 
 urlpatterns = [
-
     url('^teams/$', views.TeamListView.as_view(), name='team_list'),
     url('^teams/create$', views.TeamCreateView.as_view(), name='team_create'),
     url('^teams/(?P<pk>\d+)/$', views.TeamDetailView.as_view(), name='team_detail'),
