+++ localshop/apps/permissions/admin.py
@@ -3,16 +3,6 @@ from django.contrib import admin
 from localshop.apps.permissions import models
 
 
-class TeamMemberInline(admin.TabularInline):
-    model = models.TeamMember
-    list_display = ['user', 'role']
-
-
-class TeamAdmin(admin.ModelAdmin):
-    list_display = ['name']
-    inlines = [TeamMemberInline]
-
-
 class CidrAdmin(admin.ModelAdmin):
     list_display = ['cidr', 'label']
 
@@ -23,6 +13,5 @@ class CredentialAdmin(admin.ModelAdmin):
     list_filter = ['repository', 'allow_upload']
 
 
-admin.site.register(models.Team, TeamAdmin)
 admin.site.register(models.CIDR, CidrAdmin)
 admin.site.register(models.Credential, CredentialAdmin)
