+++ localshop/templates/accounts/team_confirm_delete.html
@@ -4,8 +4,8 @@
 {% block breadcrumbs %}
 <ul class="breadcrumb">
     <li><a href="#">localshop</a> <span class="divider">/</span></li>
-    <li><a href="{% url 'permissions:team_list' %}">{% trans "teams" %}</a> <span class="divider">/</span></li>
-    <li><a href="{% url 'permissions:team_detail' pk=team.pk %}">{{ team.name }}</a> <span class="divider">/</span></li>
+    <li><a href="{% url 'accounts:team_list' %}">{% trans "teams" %}</a> <span class="divider">/</span></li>
+    <li><a href="{% url 'accounts:team_detail' pk=team.pk %}">{{ team.name }}</a> <span class="divider">/</span></li>
     <li class="active">{% trans "delete" %}</li>
 </ul>
 {% endblock %}
@@ -18,7 +18,7 @@
 
     <div class="form-actions">
         <button type="submit" class="btn btn-primary"><span class="icon-trash icon-white"></span> Delete</button>
-        <a href="{% url 'permissions:team_detail' pk=team.pk %}" class="btn">Cancel</a>
+        <a href="{% url 'accounts:team_detail' pk=team.pk %}" class="btn">Cancel</a>
     </div>
     {% csrf_token %}
 </form>
