+++ localshop/templates/accounts/team_detail.html
@@ -4,7 +4,7 @@
 {% block breadcrumbs %}
 <ul class="breadcrumb">
     <li><a href="#">localshop</a> <span class="divider">/</span></li>
-    <li><a href="{% url 'permissions:team_list' %}">{% trans "teams" %}</a> <span class="divider">/</span></li>
+    <li><a href="{% url 'accounts:team_list' %}">{% trans "teams" %}</a> <span class="divider">/</span></li>
     <li class="active">{{ team.name }}</li>
 </ul>
 {% endblock %}
@@ -14,10 +14,10 @@
 <div>
     <h1>{{ team }}</h1>
     <p>{{ team.description }}</p>
-    <a href="{% url 'permissions:team_edit' pk=team.pk %}"><span class="icon-edit"></span> edit</a> |
-    <a href="{% url 'permissions:team_delete' pk=team.pk %}"><span class="icon-trash"></span> delete</a>
+    <a href="{% url 'accounts:team_edit' pk=team.pk %}"><span class="icon-edit"></span> edit</a> |
+    <a href="{% url 'accounts:team_delete' pk=team.pk %}"><span class="icon-trash"></span> delete</a>
 
-    <form method="post" action="{% url 'permissions:team_member_add' pk=team.pk %}" class="form form-inline">
+    <form method="post" action="{% url 'accounts:team_member_add' pk=team.pk %}" class="form form-inline">
         {% include "partial/_form_field.html" with field=form_member_add.user %}
         {% include "partial/_form_field.html" with field=form_member_add.role %}
         {% csrf_token %}
@@ -37,7 +37,7 @@
                 <td>{{ member.user }}</td>
                 <td>{{ member.role }}</td>
                 <td>
-                    <form method="post" action="{% url 'permissions:team_member_remove' pk=team.pk %}">
+                    <form method="post" action="{% url 'accounts:team_member_remove' pk=team.pk %}">
                         <input type="hidden" name="member_obj" value="{{ member.pk }}">
                         {% csrf_token %}
                         <button type="submit">X</button>
