+++ localshop/templates/layout.main.html
@@ -11,7 +11,7 @@
             <div class="nav-collapse">
                 <ul class="nav">
                     <li><a href="{% url 'dashboard:repository_list' %}">Repositories</a></li>
-                    <li><a href="{% url 'permissions:team_list' %}">Teams</a></li>
+                    <li><a href="{% url 'accounts:team_list' %}">Teams</a></li>
                 </ul>
 
                 <ul class="nav pull-right">
