+++ localshop/urls.py
@@ -22,10 +22,10 @@ urlpatterns = [
 
     url(r'^repo/',
         include('localshop.apps.packages.urls', namespace='packages')),
-    url(r'^permissions/',
-        include('localshop.apps.permissions.urls', namespace='permissions')),
 
     url(r'^accounts/', include('django.contrib.auth.urls')),
+    url(r'^accounts/',
+        include('localshop.apps.accounts.urls', namespace='accounts')),
     url(r'^admin/', include(admin.site.urls)),
 
     url(r'^%s(?P<path>.*)$' % static_prefix,
